python stream binder: run generic script and pass command line arguments
The default script to run is 'run_all_tests.py', that executes all
other python scripts on the same folder that starts with 'test_'.
The origin 'script.py' was changed to comply with this mechanism and
renamed to 'test_predifined_java_src.py'.
It is also possible to run any other desired python script by providing
a command line argument with the format:
'--script <script-name-in-resouces>'
or, just by providing the full path to the script.